% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBShiny.R
\name{NBShiny}
\alias{NBShiny}
\title{Launch 'NBShiny' Interface}
\usage{
NBShiny()
}
\value{
Nothing
}
\description{
NBShiny() loads interactive user interface built using R 'shiny'.
}
\details{
Start NBShiny

The interactive user interface is to provide an easy way for naive bayesian classification analysis and downloading relevant plot.
}
\examples{
if(interactive()){
library(rmarkdown)
NBShiny()
}
}
\keyword{NBShiny}
